+++ src/main/java/com/github/rainang/tilelib/board/Board.java
@@ -1,0 +1,158 @@
+package com.github.rainang.tilelib.board;
+
+import com.github.rainang.tilelib.board.tile.Tile;
+import com.github.rainang.tilelib.point.Point;
+
+import java.util.*;
+import java.util.function.Function;
+
+public class Board<T extends Tile>
+{
+	private final Map<Point, T> tileMap;
+	
+	Board(Set<Point> points, Function<Point, T> tileSupplier)
+	{
+		Map<Point, T> map = new HashMap<>();
+		points.forEach(c -> map.put(c, tileSupplier.apply(c)));
+		tileMap = Collections.unmodifiableMap(map);
+	}
+	
+	public Map<Point, T> getTileMap()
+	{
+		return tileMap;
+	}
+	
+	public Set<Point> getCoordinates()
+	{
+		return tileMap.keySet();
+	}
+	
+	public Collection<T> getTiles()
+	{
+		return tileMap.values();
+	}
+	
+	public T getTile(Point pos)
+	{
+		return tileMap.get(pos);
+	}
+	
+	public boolean contains(Point pos)
+	{
+		return tileMap.containsKey(pos);
+	}
+	
+	public static abstract class Builder<T extends Tile, B extends Board<T>> implements javafx.util.Builder<B>
+	{
+		private final TileFinder finder;
+		
+		private final Set<Point> points = new HashSet<>();
+		
+		Builder(TileFinder finder)
+		{
+			this.finder = finder;
+		}
+		
+		public TileFinder getFinder()
+		{
+			return finder;
+		}
+		
+		public boolean addCoordinates(Collection<Point> points)
+		{
+			return this.points.addAll(points);
+		}
+		
+		public boolean removeCoordinates(Collection<Point> points)
+		{
+			return this.points.removeAll(points);
+		}
+		
+		public boolean addCoordinate(Point point)
+		{
+			return points.add(point);
+		}
+		
+		public boolean removeCoordinate(Point point)
+		{
+			return points.remove(point);
+		}
+		
+		public Set<Point> getPoints()
+		{
+			return Collections.unmodifiableSet(points);
+		}
+		
+		public boolean validateArena()
+		{
+			List<Point> coords = new ArrayList<>(points);
+			Queue<Point> search = new ArrayDeque<>();
+			
+			search.offer(coords.get(0));
+			
+			while (!search.isEmpty())
+			{
+				Point hc = search.poll();
+				coords.remove(hc);
+				for (int i = 0; i < 6; i++)
+				{
+					Point offset = finder.offset(hc, i);
+					if (!search.contains(offset) && coords.contains(offset))
+						search.offer(offset);
+				}
+			}
+			
+			return coords.isEmpty();
+		}
+	}
+	
+	public static <T extends Tile> Board.Builder<T, Board<T>> defaultHexBuilder(Function<Point, T> tileSupplier)
+	{
+		return new Builder<T, Board<T>>(new HexFinder())
+		{
+			@Override
+			public Board<T> build()
+			{
+				return new Board<>(getPoints(), tileSupplier);
+			}
+		};
+	}
+	
+	public static <T extends Tile> Board.Builder<T, Board<T>> defaultQuadBuilder(Function<Point, T> tileSupplier)
+	{
+		return new Builder<T, Board<T>>(new QuadFinder())
+		{
+			@Override
+			public Board<T> build()
+			{
+				return new Board<>(getPoints(), tileSupplier);
+			}
+		};
+	}
+	
+	public static <B extends Board<T>, T extends Tile> Board.Builder<T, Board<T>> customHexBuilder
+			(Function<Set<Point>, B> boardSupplier)
+	{
+		return new Builder<T, Board<T>>(new HexFinder())
+		{
+			@Override
+			public B build()
+			{
+				return boardSupplier.apply(getPoints());
+			}
+		};
+	}
+	
+	public static <B extends Board<T>, T extends Tile> Board.Builder<T, Board<T>> customQuadBuilder
+			(Function<Set<Point>, B> boardSupplier)
+	{
+		return new Builder<T, Board<T>>(new QuadFinder())
+		{
+			@Override
+			public B build()
+			{
+				return boardSupplier.apply(getPoints());
+			}
+		};
+	}
+}