+++ src/main/java/com/github/rainang/tilelib/board/tile/Foggable.java
@@ -1,4 +1,4 @@
-package com.github.rainang.tilelib.tiles;
+package com.github.rainang.tilelib.board.tile;
 
 public interface Foggable
 {