+++ src/main/java/com/github/rainang/tilelib/board/tile/Weighable.java
@@ -1,4 +1,4 @@
-package com.github.rainang.tilelib.tiles;
+package com.github.rainang.tilelib.board.tile;
 
 /**
  Interface for Tile classes that can hold weight for obstructing movement