+++ src/main/java/com/github/rainang/tilelib/coordinates/CoordinateSupplier.java
@@ -1,7 +1,0 @@
-package com.github.rainang.tilelib.coordinates;
-
-@FunctionalInterface
-public interface CoordinateSupplier<C extends Coordinate>
-{
-	C create(int i, int j);
-}