+++ src/main/java/com/github/rainang/tilelib/tiles/HexTile.java
@@ -1,9 +1,0 @@
-package com.github.rainang.tilelib.tiles;
-
-import com.github.rainang.tilelib.coordinates.HexCoordinate;
-
-public interface HexTile extends Tile
-{
-	@Override
-	HexCoordinate getPos();
-}