+++ src/main/java/com/github/rainang/tilelib/tiles/HexTileMap.java
@@ -1,5 +1,0 @@
-package com.github.rainang.tilelib.tiles;
-
-import com.github.rainang.tilelib.coordinates.HexCoordinate;
-
-public class HexTileMap<T extends HexTile> extends TileMapBase<HexCoordinate, T> {}