+++ src/main/java/com/github/rainang/tilelib/tiles/Tile.java
@@ -1,8 +1,0 @@
-package com.github.rainang.tilelib.tiles;
-
-import com.github.rainang.tilelib.coordinates.Coordinate;
-
-public interface Tile
-{
-	Coordinate getPos();
-}