+++ src/main/java/com/github/rainang/tilelib/tiles/TileMap.java
@@ -1,6 +1,0 @@
-package com.github.rainang.tilelib.tiles;
-
-import com.github.rainang.tilelib.coordinates.Coordinate;
-
-public class TileMap<T extends Tile> extends TileMapBase<Coordinate, T>
-{}