+++ src/main/java/com/github/rainang/tilelib/tiles/TileMapBase.java
@@ -1,8 +1,0 @@
-package com.github.rainang.tilelib.tiles;
-
-import com.github.rainang.tilelib.coordinates.Coordinate;
-
-import java.util.HashMap;
-
-public class TileMapBase<C extends Coordinate, T extends Tile> extends HashMap<C, T>
-{}